Convert to using GNU make as the preferred build
