--- .hgignore
@@ -1,5 +1,0 @@
-^build$
-^dist$
-^nbproject$
-^\.idea$
-doccheck.iml
