+++ build.xml
@@ -12,7 +12,7 @@
         <mkdir dir="${dist.dir}"/>
         <jar destfile="${dist.dir}/doccheck.jar" basedir="${build.dir}/classes">
             <manifest>
-                <attribute name="Main-Class" value="doccheck.Main"/>
+                <attribute name="Main-Class" value="jdk.codetools.doccheck.Main"/>
             </manifest>
         </jar>
     </target>
