+++ src/classes/jdk/codetools/doccheck/HtmlFileChecker.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package doccheck;
+package jdk.codetools.doccheck;
 
 import java.io.BufferedReader;
 import java.io.IOException;
