+++ src/classes/jdk/codetools/doccheck/TextReporter.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package doccheck;
+package jdk.codetools.doccheck;
 
 import java.io.PrintWriter;
 import java.util.Iterator;
