+++ src/classes/jdk/codetools/doccheck/html/BadCharacterChecker.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package doccheck.html;
+package jdk.codetools.doccheck.html;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
@@ -39,9 +39,9 @@ import java.nio.file.Path;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import doccheck.FileChecker;
-import doccheck.Log;
-import doccheck.Reporter;
+import jdk.codetools.doccheck.FileChecker;
+import jdk.codetools.doccheck.Log;
+import jdk.codetools.doccheck.Reporter;
 
 /**
  * Checks the contents of an HTML file for bad/unmappable characters.
